band-platform/backend/requirements.txt
# FastAPI and related dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0
alembic==1.13.0

# Google APIs
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Data validation and settings
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# HTTP client
httpx==0.25.2
aiofiles==23.2.1

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
black==23.11.0
ruff==0.1.6
mypy==1.7.1

# Production
redis==5.0.1
celery==5.3.4

# CORS
fastapi-cors==0.0.6